reuse projectFilter for http2 and socket.io
